requirements.txt
Django==4.2.7
Pillow==10.0.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
django-allauth==0.63.6
requests==2.31.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
django-redis==6.0.0
redis==4.4.4
